Injecte la vue de la situation « Inventaire » dans le cadre
