+++ src/app/situationInventaire.js
@@ -6,43 +6,29 @@ import 'inventaire/styles/etageres.scss';
 import { DepotJournal } from 'commun/infra/depot_journal.js';
 import { Journal } from 'commun/modeles/journal.js';
 import { ActionsCommunesSituation } from 'commun/vues/actions_communes_situation.js';
+import { VueCadre } from 'commun/vues/cadre.js';
 import { VueConsigne } from 'commun/vues/consigne.js';
 import { VueGo } from 'commun/vues/go.js';
-import { contenants, contenus } from 'inventaire/data/stock.js';
-import { creeMagasin } from 'inventaire/modeles/magasin.js';
-import { VueEtageres } from 'inventaire/vues/etageres.js';
-import { afficheCorrection, initialiseFormulaireSaisieInventaire } from 'inventaire/vues/formulaireSaisieInventaire.js';
+import { VueSituation } from 'inventaire/vues/situation.js';
 
 import sonConsigneDemarrage from 'inventaire/assets/consigne_demarrage.mp3';
 
-function afficheMagasin (pointInsertion, $) {
+function afficheSituation (pointInsertion, $) {
   const session = uuidv4();
-  const magasin = creeMagasin({ contenants, contenus });
   const journal = new Journal(Date.now, session, new DepotJournal());
-  new VueEtageres(pointInsertion, journal)
-    .affiche(magasin.contenants());
+  const vueSituationInventaire = new VueSituation(journal);
+  const vueCadre = new VueCadre(vueSituationInventaire);
+  vueCadre.affiche(pointInsertion, $);
 
-  initialiseFormulaireSaisieInventaire(magasin, pointInsertion, $, function (resultatValidation, reponses) {
-    const toutCorrect = Array.from(resultatValidation.values()).every(v => v);
-    const message = toutCorrect ? 'Bravo, vous avez réussi !' : 'Ce n\'est pas tout à fait ça… réessayez.';
-
-    journal.enregistreSaisieInventaire(toutCorrect, reponses);
-
-    Array.from(resultatValidation).forEach((correction) => { afficheCorrection(correction, $); });
-    window.alert(message);
-  }, journal);
   new ActionsCommunesSituation(pointInsertion, $, journal).afficheElementEnCommun();
+
   const vueConsigne = new VueConsigne(pointInsertion, sonConsigneDemarrage);
   new VueGo(pointInsertion, vueConsigne, journal).afficher();
 }
 
 jQuery(function () {
   document.title = 'Situation Inventaire';
-  jQuery('body').append(`
-    <div id="magasin" class="situation">
-      <h1>Stock de boissons</h1>
-    </div>
-    `);
+  jQuery('body').append(`<div id="magasin"> </div>`);
 
-  afficheMagasin('#magasin', jQuery);
+  afficheSituation('#magasin', jQuery);
 });
