+++ src/situations/commun/styles/commun.scss
@@ -7,13 +7,3 @@ ul {
   margin: 0;
   padding: 0;
 }
-
-.situation {
-  position: relative;
-  margin: 2rem auto 0;
-  width: 60rem;
-  height: 40rem;
-  border: solid 1px black;
-  background: ivory;
-  overflow: hidden;
-}
