+++ src/situations/inventaire/styles/etageres.scss
@@ -1,7 +1,6 @@
 .etageres {
   position: relative;
-  height: 86%;
-  margin: 2%;
+  height: 100%;
 
   .avant-plan {
     position:absolute;
